Create a breadcrumb VSIX that Visual FSharp Tools can use to notify F# users to download the Visual F# Power Tools.
Vsix is built to the ..\bin directory, it is called FSharpVSPowerTools.Breadcrumb.vsix

To test:
	Build FSharpVSPowerTools.sln
	Ensure that the powerstools are not installed.
	Ensure VS is not running
	Install FSharpVSPowerTools.Breadcrumb.vsix from the build directory
	Start VS and wait a minute or two, you should see a little update notification flag in the top left of the screen
	Click on the flag in the list of updates you should notice the Visual FSharp Power Tools.

If you want to submit this as a pull request to Visual FSharp Tools, we will see if we can include it in the fsharp Oob.
That should make it much easier for VS users to get at the power tools.

Kevin Ransom