add empty todo tests.
need to refresh the state of the tokenizer when scanning the line
comment, if become the new line.
because scanning of line comments may be canceled in the middle.
scan of the entire line is not required for the detection of the task
list.

// this is the line comment and state must be kept until the end of line
but_this_ident_must_be_scaned_by_the_state_before_line_comment