Commit because F# powertools broke the project file
Tried to add a file inside of a folder, it didn't get added to the
correct folder, closed and reopened the solution and the project
wouldn't load

" error  : The project 'FSharpVSPowerTools.Logic.fsproj' could not be
opened because opening it would cause a folder to be rendered multiple
times in the solution explorer. One such problematic item is
'Outlining\LineRanges.fs'. To open this project in Visual Studio, first
edit the project file and fix the problem."